Move several modules into a common sub-package, fix imports that reference them
