+++ src/vak/cli/train.py
@@ -4,8 +4,8 @@ import shutil
 
 from .. import config
 from .. import train as train_module
-from ..logging import config_logging_for_cli, log_version
-from ..paths import generate_results_dir_name_as_path
+from ..common.logging import config_logging_for_cli, log_version
+from ..common.paths import generate_results_dir_name_as_path
 
 
 logger = logging.getLogger(__name__)
