+++ src/vak/common/__init__.py
@@ -1,0 +1,24 @@
+"""The :mod:`vak.common` module contains helper or utility functions
+that are used in multiple other modules, e.g., validators
+or logging functions.
+
+If a helper/utility function is only used in one module,
+it should live either in that module or another at the same level.
+See for example :mod:`vak.prep.prep_helper` or
+:mod:`vak.datsets.window_dataset._helper`.
+"""
+from . import (
+    annotation,
+    constants,
+    converters,
+    device,
+    labels,
+    logging,
+    paths,
+    tensorboard,
+    timebins,
+    timenow,
+    trainer,
+    typing,
+    validators,
+)
