+++ src/vak/common/annotation.py
