+++ src/vak/common/device.py
