+++ src/vak/common/timebins.py
