+++ src/vak/common/timenow.py
