+++ src/vak/common/trainer.py
