+++ src/vak/common/typing.py
