+++ src/vak/config/parse.py
@@ -13,6 +13,7 @@ from .spect_params import SpectParamsConfig
 from .train import TrainConfig
 from .validators import are_sections_valid, are_options_valid
 
+
 SECTION_CLASSES = {
     "DATALOADER": DataLoaderConfig,
     "EVAL": EvalConfig,
