+++ src/vak/config/predict.py
@@ -8,7 +8,7 @@ from attr.validators import instance_of
 
 from .validators import is_valid_model_name
 from .. import device
-from ..converters import expanded_user_path
+from ..common.converters import expanded_user_path
 
 
 @attr.s
