+++ src/vak/datasets/seq/validators.py
@@ -7,7 +7,7 @@ import numpy as np
 import numpy.typing as npt
 import pandas as pd
 
-from ... import annotation, files
+from ...common import annotation, files
 
 
 def where_unlabeled(dataset_csv_path: str | pathlib.Path,
