+++ src/vak/prep/prep.py
@@ -9,11 +9,11 @@ from . import prep_helper
 from .learncurve import make_learncurve_splits_from_dataset_df
 
 from .. import datasets, labels, split
-from ..converters import expanded_user_path, labelset_to_set
+from ..common.converters import expanded_user_path, labelset_to_set
 from ..datasets.metadata import Metadata
 from ..io import dataframe
-from ..logging import config_logging_for_cli, log_version
-from ..timenow import get_timenow_as_str
+from ..common.logging import config_logging_for_cli, log_version
+from ..common.timenow import get_timenow_as_str
 
 
 __all__ = [
