+++ src/vak/transforms/transforms.py
@@ -4,9 +4,10 @@ import numpy as np
 import pandas as pd
 
 from .. import files
-from ..validators import column_or_1d
+from ..common.validators import column_or_1d
 
 from . import functional as F
+
 
 __all__ = [
     "AddChannel",
